refactored back-end validation add atlas props to test profile
