+++ src/main/java/com/odysseusinc/arachne/datanode/controller/ExceptionHandlingController.java
@@ -20,16 +20,16 @@
  *
  */
 
-package com.odysseusinc.arachne.datanode.config;
+package com.odysseusinc.arachne.datanode.controller;
 
 import static com.odysseusinc.arachne.commons.api.v1.dto.util.JsonResult.ErrorCode.SYSTEM_ERROR;
 import static com.odysseusinc.arachne.commons.api.v1.dto.util.JsonResult.ErrorCode.UNAUTHORIZED;
 import static com.odysseusinc.arachne.commons.api.v1.dto.util.JsonResult.ErrorCode.VALIDATION_ERROR;
 
 import com.odysseusinc.arachne.commons.api.v1.dto.util.JsonResult;
-import com.odysseusinc.arachne.datanode.controller.BaseController;
 import com.odysseusinc.arachne.datanode.exception.AuthException;
 import com.odysseusinc.arachne.datanode.exception.IntegrationValidationException;
+import com.odysseusinc.arachne.datanode.exception.NotExistException;
 import com.odysseusinc.arachne.datanode.exception.ValidationException;
 import com.odysseusinc.arachne.datanode.service.UserService;
 import java.io.IOException;
@@ -44,18 +44,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestController;
 
-@RestController
 @ControllerAdvice
-public class ExceptionHandlingAdvice extends BaseController {
+public class ExceptionHandlingController extends BaseController {
 
     private static final String ERROR_MESSAGE = "Please contact system administrator and provide this error token: %s";
-    private static final Logger LOGGER = LoggerFactory.getLogger(ExceptionHandlingAdvice.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(ExceptionHandlingController.class);
     @Value("${datanode.app.errorsTokenEnabled}")
     private boolean errorsTokenEnabled;
 
-    public ExceptionHandlingAdvice(UserService userService) {
+    public ExceptionHandlingController(UserService userService) {
 
         super(userService);
     }
@@ -136,6 +134,14 @@ public class ExceptionHandlingAdvice extends BaseController {
 
         LOGGER.error(ex.getMessage(), ex);
         JsonResult result = new JsonResult<>(VALIDATION_ERROR);
+        result.setErrorMessage(ex.getMessage());
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
+
+    @ExceptionHandler(NotExistException.class)
+    public ResponseEntity<JsonResult> exceptionHandler(NotExistException ex) {
+        LOGGER.error(ex.getMessage());
+        JsonResult result = new JsonResult<>(SYSTEM_ERROR);
         result.setErrorMessage(ex.getMessage());
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
