+++ src/test/resources/config/application.yml
@@ -59,5 +59,14 @@ docker:
   corePoolSize: 5
   maxPoolSize: 5
   queueCapacity: 20
+atlas:
+  host: https://testatlas.arachnenetwork.com
+  port: 443
+  auth:
+    schema: DATABASE
+    username: admin@odysseusinc.com
+    password: password
+  scheduler:
+    checkInterval: 10000
 server:
   port: 8080
